# 🚀 Massively speed up fetching of image dimensions