--- a/lib/amperize.js
+++ b/lib/amperize.js
@@ -7,7 +7,8 @@ var EventEmitter = require('events').EventEmitter,
     uuid = require('uuid'),
     async = require('async'),
     url = require('url'),
-    got = require('got'),
+    request = require('request-promise'),
+    probeImageSize = require('probe-image-size'),
     _ = require('lodash'),
     sizeOf = require('image-size'),
     validator = require('validator'),
@@ -105,6 +106,16 @@ Amperize.prototype.amperizer = function amperizer(id, error, dom) {
 */
 Amperize.prototype.traverse = function traverse(data, html, done) {
     var self = this;
+    var imageSizeCache = {};
+    var timeout = 3000;
+    var requestOptions = {
+        // We need the user-agent, otherwise some https request may fail (e. g. cloudfare)
+        headers: {
+            'User-Agent': 'Mozilla/5.0 Safari/537.36'
+        },
+        timeout: timeout,
+        encoding: null
+    };
 
     async.reduce(data, html, function reduce(html, element, step) {
         var children;
@@ -158,62 +169,86 @@ Amperize.prototype.traverse = function traverse(data, html, done) {
             return enter();
         }
 
+        // probe will fetch the minimal amount of data needed to determine
+        // the image dimensions so it's more performant than a full fetch
+        function _probeImageSize(url) {
+            return probeImageSize(
+                url,
+                requestOptions
+            ).then(function (result) {
+                imageSizeCache[url] = result;
+                return result;
+            });
+        }
+
+        // fetch the full image before reading dimensions using image-size,
+        // it's slower but has better format support
+        function _fetchImageSize(url) {
+            return request(
+                url,
+                Object.assign({}, requestOptions, {
+                    encoding: null
+                })
+            ).then(function (response) {
+                var result = sizeOf(response);
+                imageSizeCache[url] = result;
+                return result;
+            });
+        }
+
+        // select appropriate method to get image size
+        function _getImageSize(url) {
+            var [, extension] = url.match(/(?:\.)([a-zA-Z]{3,4})$/) || [];
+
+            // use cached image size if we've already seen this url
+            if (imageSizeCache[url]) {
+                return Promise.resolve(imageSizeCache[url]);
+            }
+
+            // // fetch full image for formats we can't probe 
+            if (['cur', 'icns', 'ico', 'dds'].includes(extension)) {
+                return _fetchImageSize(url);
+            }
+            
+            // // probe partial image everything else
+            return _probeImageSize(url);
+        }
+
         /**
-        * Get the image sizes (width and heigth plus type of image)
+        * Get the image sizes (width and height plus type of image)
         *
-        * https://github.com/image-size/image-size
+        * https://github.com/nodeca/probe-image-size
         *
         * @param {Object} element
         * @return {Object} element incl. width and height
         */
         function getImageSize(element) {
-            var imageObj = url.parse(element.attribs.src),
-                requestOptions,
-                timeout = 3000;
+            var imageObj = url.parse(element.attribs.src);
 
             if (!validator.isURL(imageObj.href)) {
                 // revert this element, do not show
                 element.name = 'img';
-
                 return enter();
             }
 
-            // We need the user-agent, otherwise some https request may fail (e. g. cloudfare)
-            requestOptions = {
-                headers: {
-                    'User-Agent': 'Mozilla/5.0 Safari/537.36'
-                },
-                timeout: timeout,
-                retry: 0,
-                encoding: null
-            };
-
-            return got(
-                imageObj.href,
-                requestOptions
-            ).then(function (response) {
-                try {
-                    // Using the Buffer rather than an URL requires to use sizeOf synchronously.
-                    // See https://github.com/image-size/image-size#asynchronous
-                    var dimensions = sizeOf(response.body);
-
-                    // CASE: `.ico` files might have multiple images and therefore multiple sizes.
-                    // We return the largest size found (image-size default is the first size found)
-                    if (dimensions.images) {
-                        dimensions.width = _.maxBy(dimensions.images, function (w) {return w.width;}).width;
-                        dimensions.height = _.maxBy(dimensions.images, function (h) {return h.height;}).height;
-                    }
-
-                    element.attribs.width = dimensions.width;
-                    element.attribs.height = dimensions.height;
-
-                    return getLayoutAttribute(element);
-                } catch (err) {
-                    // revert this element, do not show
+            return _getImageSize(imageObj.href).then(function (result) {
+                if ((!result.width || !result.height) && !result.images) {
                     element.name = 'img';
                     return enter();
                 }
-            }).catch(function () {
+
+                // CASE: `.ico` files might have multiple images and therefore multiple sizes.
+                // We return the largest size found (image-size default is the first size found)
+                if (result.images) {
+                    result.width = _.maxBy(result.images, function (w) {return w.width;}).width;
+                    result.height = _.maxBy(result.images, function (h) {return h.height;}).height;
+                }
+
+                element.attribs.width = result.width;  
+                element.attribs.height = result.height;
+
+                return getLayoutAttribute(element);
+            }).catch(function (err) {
                 // revert this element, do not show
                 element.name = 'img';
                 return enter();