--- a/lib/amperize.js
+++ b/lib/amperize.js
@@ -3,6 +3,7 @@
 var EventEmitter = require('events').EventEmitter,
     emits = require('emits'),
     html = require('htmlparser2'),
+    domutils = require('domutils'),
     util = require('util'),
     uuid = require('uuid'),
     async = require('async'),
@@ -104,19 +105,154 @@ Amperize.prototype.amperizer = function amperizer(id, error, dom) {
 * @param {Function} done Callback function
 * @api private
 */
-Amperize.prototype.traverse = function traverse(data, html, done) {
+Amperize.prototype.traverse = async function traverse(data, html, done) {
     var self = this;
     var imageSizeCache = {};
-    var timeout = 3000;
+
     var requestOptions = {
         // We need the user-agent, otherwise some https request may fail (e. g. cloudfare)
         headers: {
             'User-Agent': 'Mozilla/5.0 Safari/537.36'
         },
-        timeout: timeout,
+        timeout: 3000,
         encoding: null
     };
 
+    // check if element.width is smaller than 300 px. In that case, we shouldn't use
+    // layout="responsive", because the media element will be stretched and it doesn't
+    // look nice. Use layout="fixed" instead to fix that.
+    function setLayoutAttribute(element) {
+        var layout = element.attribs.width < 300 ? layout = 'fixed' : self.config[element.name].layout;
+        element.attribs.layout = !element.attribs.layout ? layout : element.attribs.layout;
+    }
+
+    // Certain component src attribute must be with 'https' protocol otherwise it will not
+    // get validated by AMP. If we're unable to replace it, we will deal with the valitation
+    // error, but at least we tried.
+    function useSecureSchema(element) {
+        if (element.attribs && element.attribs.src) {
+            if (element.attribs.src.indexOf('https://') === -1) {
+                if (element.attribs.src.indexOf('http://') === 0) {
+                    // Replace 'http' with 'https', so the validation passes
+                    element.attribs.src = element.attribs.src.replace(/^http:\/\//i, 'https://');
+                } else if (element.attribs.src.indexOf('//') === 0) {
+                    // Giphy embedded iFrames are without protocol and start with '//', so at least
+                    // we can fix those cases.
+                    element.attribs.src = 'https:' + element.attribs.src;
+                }
+            }
+        }
+    }
+
+    // probe will fetch the minimal amount of data needed to determine
+    // the image dimensions so it's more performant than a full fetch
+    function _probeImageSize(url) {
+        return probeImageSize(
+            url,
+            requestOptions
+        ).then(function (result) {
+            imageSizeCache[url] = result;
+            return result;
+        });
+    }
+
+    // fetch the full image before reading dimensions using image-size,
+    // it's slower but has better format support
+    function _fetchImageSize(url) {
+        return request(
+            url,
+            requestOptions
+        ).then(function (response) {
+            var result = sizeOf(response);
+            imageSizeCache[url] = result;
+            return result;
+        });
+    }
+
+    // select appropriate method to get image size
+    function _getImageSize(url) {
+        // use cached image size if we've already seen this url
+        if (imageSizeCache[url]) {
+            return Promise.resolve(imageSizeCache[url]);
+        }
+
+        var [, extension] = url.match(/(?:\.)([a-zA-Z]{3,4})$/) || [];
+
+        // fetch full image for formats we can't probe 
+        if (['cur', 'icns', 'ico', 'dds'].includes(extension)) {
+            return _fetchImageSize(url);
+        }
+        
+        // probe partial image everything else
+        return _probeImageSize(url);
+    }
+
+    // convert <img> to <amp-img> or <amp-anim>, fetching dimensions of
+    // external images. If anything fails leave the element as an <img>
+    function amperizeImageElem(element) {
+        return async function() {
+            if (!element.attribs || !element.attribs.src) {
+                return;
+            }
+
+            var src = url.parse(element.attribs.src).href;
+
+            // when we have a gif it should be <amp-anim>.
+            element.name = src.match(/(\.gif$)/) ? 'amp-anim' : 'amp-img';
+
+            if (src.indexOf('http') === 0) {
+                // external image, fetch real dimensions
+                try {
+                    if (!validator.isURL(src)) {
+                        element.name = 'img';
+                        return;
+                    }
+
+                    var dimensions = await _getImageSize(src);
+
+                    // CASE: `.ico` files might have multiple images and therefore multiple sizes.
+                    // We return the largest size found (image-size default is the first size found)
+                    if (dimensions.images) {
+                        dimensions.width = _.maxBy(dimensions.images, function (w) {return w.width;}).width;
+                        dimensions.height = _.maxBy(dimensions.images, function (h) {return h.height;}).height;
+                    }
+
+                    if (!dimensions.width || !dimensions.height) {
+                        element.name = 'img';
+                        return;
+                    }
+
+                    element.attribs.width = dimensions.width;  
+                    element.attribs.height = dimensions.height;
+
+                } catch (err) {
+                    element.name = 'img';
+                    return;
+                }
+            } else {
+                // local image, use default fallback
+                element.attribs.width = self.config[element.name].width;
+                element.attribs.height = self.config[element.name].height;
+            }
+
+            if (!element.attribs.layout) {
+                setLayoutAttribute(element);
+            }
+        }
+    }
+
+    // convert all of the img elements first so that we can perform lengthy
+    // network requests in parallel before sequentially traversing the DOM
+    if (self.config['amp-img']) {
+        var imgTest = function(elem) {
+            return elem.name === 'img' && elem.attribs.src;
+        }
+        var imgElems = domutils.findAll(elem => imgTest(elem), data);
+        var imgTasks = imgElems.map(elem => amperizeImageElem(elem));
+        await async.parallelLimit(imgTasks, 10);
+    }
+
+    // sequentially traverse the DOM
     async.reduce(data, html, function reduce(html, element, step) {
         var children;
 
@@ -138,161 +274,30 @@ Amperize.prototype.traverse = function traverse(data, html, done) {
             });
         }
 
-        function useSecureSchema(element) {
-            if (element.attribs && element.attribs.src) {
-                // Every src attribute must be with 'https' protocol otherwise it will not get validated by AMP.
-                // If we're unable to replace it, we will deal with the valitation error, but at least
-                // we tried.
-                if (element.attribs.src.indexOf('https://') === -1) {
-                    if (element.attribs.src.indexOf('http://') === 0) {
-                        // Replace 'http' with 'https', so the validation passes
-                        element.attribs.src = element.attribs.src.replace(/^http:\/\//i, 'https://');
-                    } else if (element.attribs.src.indexOf('//') === 0) {
-                        // Giphy embedded iFrames are without protocol and start with '//', so at least
-                        // we can fix those cases.
-                        element.attribs.src = 'https:' + element.attribs.src;
-                    }
-                }
-            }
-        }
-
-        function getLayoutAttribute(element) {
-            var layout;
-
-            // check if element.width is smaller than 300 px. In that case, we shouldn't use
-            // layout="responsive", because the media element will be stretched and it doesn't
-            // look nice. Use layout="fixed" instead to fix that.
-            layout = element.attribs.width < 300 ? layout = 'fixed' : self.config[element.name].layout;
-
-            element.attribs.layout = !element.attribs.layout ? layout : element.attribs.layout;
-
-            return enter();
-        }
-
-        // probe will fetch the minimal amount of data needed to determine
-        // the image dimensions so it's more performant than a full fetch
-        function _probeImageSize(url) {
-            return probeImageSize(
-                url,
-                requestOptions
-            ).then(function (result) {
-                imageSizeCache[url] = result;
-                return result;
-            });
-        }
-
-        // fetch the full image before reading dimensions using image-size,
-        // it's slower but has better format support
-        function _fetchImageSize(url) {
-            return request(
-                url,
-                Object.assign({}, requestOptions, {
-                    encoding: null
-                })
-            ).then(function (response) {
-                var result = sizeOf(response);
-                imageSizeCache[url] = result;
-                return result;
-            });
-        }
-
-        // select appropriate method to get image size
-        function _getImageSize(url) {
-            var [, extension] = url.match(/(?:\.)([a-zA-Z]{3,4})$/) || [];
-
-            // use cached image size if we've already seen this url
-            if (imageSizeCache[url]) {
-                return Promise.resolve(imageSizeCache[url]);
-            }
-
-            // // fetch full image for formats we can't probe 
-            if (['cur', 'icns', 'ico', 'dds'].includes(extension)) {
-                return _fetchImageSize(url);
-            }
-            
-            // // probe partial image everything else
-            return _probeImageSize(url);
-        }
-
-        /**
-        * Get the image sizes (width and height plus type of image)
-        *
-        * https://github.com/nodeca/probe-image-size
-        *
-        * @param {Object} element
-        * @return {Object} element incl. width and height
-        */
-        function getImageSize(element) {
-            var imageObj = url.parse(element.attribs.src);
-
-            if (!validator.isURL(imageObj.href)) {
-                // revert this element, do not show
-                element.name = 'img';
-                return enter();
-            }
-
-            return _getImageSize(imageObj.href).then(function (result) {
-                if ((!result.width || !result.height) && !result.images) {
-                    element.name = 'img';
-                    return enter();
-                }
-
-                // CASE: `.ico` files might have multiple images and therefore multiple sizes.
-                // We return the largest size found (image-size default is the first size found)
-                if (result.images) {
-                    result.width = _.maxBy(result.images, function (w) {return w.width;}).width;
-                    result.height = _.maxBy(result.images, function (h) {return h.height;}).height;
-                }
-
-                element.attribs.width = result.width;  
-                element.attribs.height = result.height;
-
-                return getLayoutAttribute(element);
-            }).catch(function (err) {
-                // revert this element, do not show
-                element.name = 'img';
-                return enter();
-            });
-        }
-
-        if ((element.name === 'img' || element.name === 'iframe') && !element.attribs.src) {
-            return enter();
-        }
-
-        if (element.name === 'img' && self.config['amp-img']) {
-            // when we have a gif it should be <amp-anim>.
-            element.name = element.attribs.src.match(/(\.gif$)/) ? 'amp-anim' : 'amp-img';
-
-            if (!element.attribs.width || !element.attribs.height || !element.attribs.layout) {
-                if (element.attribs.src.indexOf('http') === 0) {
-                    return getImageSize(element);
-                }
-            }
-            // Fallback to default values for a local image
-            element.attribs.width = self.config['amp-img'].width;
-            element.attribs.height = self.config['amp-img'].height;
-            return getLayoutAttribute(element);
-        }
-
         if (element.name === 'iframe') {
+            if (!element.attribs.src) {
+                return enter();
+            }
+
             element.name = 'amp-iframe';
+            useSecureSchema(element);
 
             if (!element.attribs.width || !element.attribs.height || !element.attribs.layout) {
                 element.attribs.width = !element.attribs.width ? self.config['amp-iframe'].width : element.attribs.width;
                 element.attribs.height = !element.attribs.height ? self.config['amp-iframe'].height : element.attribs.height;
                 element.attribs.sandbox = !element.attribs.sandbox ? self.config['amp-iframe'].sandbox : element.attribs.sandbox;
-
-                useSecureSchema(element);
-
-                return getLayoutAttribute(element);
+                setLayoutAttribute(element);
             }
         }
 
         if (element.name === 'audio') {
             element.name = 'amp-audio';
+            useSecureSchema(element);
         }
 
-        useSecureSchema(element);
+        if (element.name === 'source' && element.parent.name === 'amp-audio') {
+            useSecureSchema(element);
+        }
 
         return enter();
     }, done);