--- a/package.json
+++ b/package.json
@@ -30,10 +30,12 @@
   "dependencies": {
     "async": "^2.1.4",
     "emits": "^3.0.0",
-    "got": "^9.6.0",
     "htmlparser2": "^3.9.2",
-    "image-size": "0.6.1",
+    "image-size": "^0.7.4",
     "lodash": "^4.17.4",
+    "probe-image-size": "^4.0.0",
+    "request": "^2.83.0",
+    "request-promise": "^4.2.4",
     "uuid": "^3.0.0",
     "validator": "^9.1.1"
   },
@@ -42,11 +44,11 @@
     "cz-conventional-changelog": "2.1.0",
     "istanbul": "^0.4.5",
     "mocha": "^4.0.1",
+    "nock": "^9.0.2",
+    "rewire": "^2.5.2",
     "semantic-release": "9.1.0",
     "sinon": "1.17.7",
-    "sinon-chai": "^2.8.0",
-    "nock": "^9.0.2",
-    "rewire": "^2.5.2"
+    "sinon-chai": "^2.8.0"
   },
   "config": {
     "commitizen": {