--- a/package.json
+++ b/package.json
@@ -28,7 +28,8 @@
   },
   "homepage": "https://github.com/jbhannah/amperize#readme",
   "dependencies": {
-    "async": "^2.1.4",
+    "async": "^3.0.1",
+    "domutils": "^1.7.0",
     "emits": "^3.0.0",
     "htmlparser2": "^3.9.2",
     "image-size": "^0.7.4",