… test/fixtures/multi-size.ico