no issue

Fetching images across the network to get their dimensions is really slow. In the previous implementation we were fetching the entire image before using it to determine the image dimensions. However, for most images we don't need all of the data! Image dimensions can usually be determined from the headers or other information that is stored near the beginning of an image file, we should take advantage of that to massively speed up our dimension calculations and save a lot of network bandwidth.

In addition to partially fetching files, this PR also addresses the sequential nature of the DOM traversal and associated network requests. By pulling img tag handling out of the traversal we're able to run those tasks in parallel as a preliminary parsing step - this results in much better utilisation of the available network capacity for another big performance increase.

- use `probe-image-size` to fetch only enough of an image that is needed to get the dimensions
- fall back to `image-size` for formats which `probe-image-size` doesn't support
- replace `got` with `request` so we're not including two different request libraries
- use `request-promise` wrapper library to simplify the code (uses `bluebird` under the hood which is the same as the `async` library that we're using
- move img tag handling out of the sequential traversal loop
  - use `domutils` to pull out an array of img elements for processing
  - create an async task for processing each img element
  - use `async.parallelLimit` to process 10 img tasks at a time

As an example of the speedup offered, here are my tests which were using this library inside Ghost. The test page had 10 reasonably sized external images and the reported time is the initial amperization before caching.

```
image-size only:
INFO amp.parse http://ghost.blog/2019/06/08/test/ 52278ms
INFO amp.parse http://ghost.blog/2019/06/08/test/ 52717ms
INFO amp.parse http://ghost.blog/2019/06/08/test/ 50582ms

average: 51,859ms

probe-image-size w/ image-size fallback:
INFO amp.parse http://ghost.blog/2019/06/08/test/ 11147ms
INFO amp.parse http://ghost.blog/2019/06/08/test/ 12297ms
INFO amp.parse http://ghost.blog/2019/06/08/test/ 11188ms

average: 11,544ms
speedup: ~4.5x

parallel image fetch before traversal:
INFO amp.parse http://ghost.blog/2019/06/08/test/ 1629ms
INFO amp.parse http://ghost.blog/2019/06/08/test/ 1744ms
INFO amp.parse http://ghost.blog/2019/06/08/test/ 1398ms

average: 1,590ms
speedup: ~7.2x
```

**Result: ~32.5x faster** 🚀