@AileenCGN 2019-06-10
@kevinansfield please submit a release PR, so we can publish the new speedy version.